+++ src/main/java/org/myworkflows/view/component/editor/IAceWordCompleter.java
@@ -1,6 +1,0 @@
-package org.myworkflows.view.component.editor;
-
-public interface IAceWordCompleter {
-
-    String toJson();
-}